KTX file format support
This commit introduces support for the KTX format in blender.

Saving and loading of files is supported though saving compressed formats
is not yet supported.

Thumbnails are not generated because the API relies on OpenGL to load the textures,
which is not available during threaded thumbnail generation.

We hand compiled a KTX library from https://github.com/KhronosGroup/KTX
and placed KTX lib files in intern/KTX as follows:

libktx.a in intern/KTX/lib
ktx.h in intern/KTX/include

We may include the library in extern when we make sure it compiles in all platforms (Ubuntu 15.04 failed here)

Scons is not supported, only cmake for now.

+++ source/blender/blenkernel/BKE_image.h
@@ -79,7 +79,7 @@ void    BKE_image_path_from_imtype(
         const char imtype, const bool use_ext, const bool use_frames, const char *suffix);
 int     BKE_image_path_ensure_ext_from_imformat(char *string, const struct ImageFormatData *im_format);
 int     BKE_image_path_ensure_ext_from_imtype(char *string, const char imtype);
-char    BKE_image_ftype_to_imtype(const int ftype);
+char    BKE_image_ftype_to_imtype(const long long int ftype);
 int     BKE_image_imtype_to_ftype(const char imtype);
 
 bool    BKE_imtype_is_movie(const char imtype);
